% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/trb.readcasecount.R
\name{trb.readcasecount}
\alias{trb.readcasecount}
\alias{trb.readcasecount}
\title{Tanrabad casecount reader}
\usage{
trb.readcasecount(txt)

trb.readcasecount(txt)
}
\arguments{
\item{txt}{casecount json text}

\item{txt}{casecount json text}
}
\value{
data frame table

data frame table
}
\description{
Tanrabad casecount reader

Tanrabad casecount reader
}
\examples{
library(Rbigstream)
host <- "http://sample.bigstream.io/"
storage_name <- "sample.trb"
token <- "token"
bs.connect(host, storage_name, token)
json <- storage.get(storage_name,flatten=FALSE)
df <- trb.readcasecount(json)
library(Rbigstream)
host <- "http://sample.bigstream.io/"
storage_name <- "sample.trb"
token <- "token"
bs.connect(host, storage_name, token)
json <- storage.get(storage_name,flatten=FALSE)
df <- trb.readcasecount(json)
}
